backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
aiofiles==24.1.0

# ── AI — Gemini 2.5 Flash ─────────────────────────────────────────────────────
# FIXED: was google-genai==1.0.0 (wrong SDK). Code uses google-generativeai API style.
google-generativeai==0.8.3

# ── Data / ML ──────────────────────────────────────────────────────────────────
pandas==2.2.3
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.2
openpyxl==3.1.5

# ── Analytics package (Member 2) ───────────────────────────────────────────────
# FIXED: replaced fuzzywuzzy with rapidfuzz (faster, no Levenshtein C dep)
rapidfuzz==3.9.7
datasketch==1.6.5
# Levenshtein only needed for detect_semantic_duplicates (optional feature)
# python-Levenshtein==0.25.1

# ── Pydantic / config ──────────────────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ── HTTP client (for testing) ──────────────────────────────────────────────────
httpx==0.27.2

# ── NLP (optional — spaCy model must be downloaded separately) ─────────────────
# spacy==3.8.2
# After install run: python -m spacy download en_core_web_sm
